Extract custom hooks and shared utilities for modular code
