+++ src/components/bryntum/hooks/useGanttControls.ts
@@ -1,0 +1,57 @@
+'use client';
+
+import { useRef, useCallback } from 'react';
+import type { BryntumGantt } from '@bryntum/gantt-react';
+
+export function useGanttControls() {
+  const ganttRef = useRef<BryntumGantt>(null);
+
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  const getGanttInstance = useCallback((): any => {
+    return (ganttRef.current as unknown as { instance: unknown })?.instance;
+  }, []);
+
+  const handleAddTask = useCallback(async () => {
+    const gantt = getGanttInstance();
+    if (!gantt) return;
+    gantt.taskStore.add({
+      name: 'New Task',
+      startDate: new Date(),
+      duration: 1,
+    });
+    await gantt.project.commitAsync();
+  }, [getGanttInstance]);
+
+  const handleZoomIn = useCallback(() => getGanttInstance()?.zoomIn(), [getGanttInstance]);
+  const handleZoomOut = useCallback(() => getGanttInstance()?.zoomOut(), [getGanttInstance]);
+  const handleZoomToFit = useCallback(() => {
+    getGanttInstance()?.zoomToFit({ leftMargin: 50, rightMargin: 50 });
+  }, [getGanttInstance]);
+  const handleShiftPrevious = useCallback(
+    () => getGanttInstance()?.shiftPrevious(),
+    [getGanttInstance]
+  );
+  const handleShiftNext = useCallback(
+    () => getGanttInstance()?.shiftNext(),
+    [getGanttInstance]
+  );
+  const handlePresetChange = useCallback(
+    (preset: string) => {
+      const gantt = getGanttInstance();
+      if (gantt) gantt.viewPreset = preset;
+    },
+    [getGanttInstance]
+  );
+
+  return {
+    ganttRef,
+    getGanttInstance,
+    handleAddTask,
+    handleZoomIn,
+    handleZoomOut,
+    handleZoomToFit,
+    handleShiftPrevious,
+    handleShiftNext,
+    handlePresetChange,
+  };
+}
