+++ src/lib/utils/formatting.ts
@@ -1,0 +1,20 @@
+/**
+ * Converts a snake_case role string to Title Case.
+ * e.g. "project_manager" → "Project Manager"
+ */
+export function formatRole(role: string): string {
+  return role
+    .split('_')
+    .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
+    .join(' ');
+}
+
+/**
+ * Formats a byte count into a human-readable file size string.
+ * e.g. 1536 → "1.5 KB"
+ */
+export function formatFileSize(bytes: number): string {
+  if (bytes < 1024) return `${bytes} B`;
+  if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`;
+  return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
+}
